requirements.txt
streamlit
PyPDF2
google-generativeai
google-cloud-documentai
python-dotenv
Pillow
pytesseract